requirements.txt
google-generativeai
streamlit
python_dotenv

